Module aws_sdk_sns::error

Modules

authorization_error_exception

See AuthorizationErrorException

concurrent_access_exception

See ConcurrentAccessException

endpoint_disabled_exception

See EndpointDisabledException

filter_policy_limit_exceeded_exception

See FilterPolicyLimitExceededException

internal_error_exception

See InternalErrorException

invalid_parameter_exception

See InvalidParameterException

invalid_parameter_value_exception

See InvalidParameterValueException

invalid_security_exception

See InvalidSecurityException

kms_access_denied_exception

See KmsAccessDeniedException

kms_disabled_exception

See KmsDisabledException

kms_invalid_state_exception

See KmsInvalidStateException

kms_not_found_exception

See KmsNotFoundException

kms_opt_in_required

See KmsOptInRequired

kms_throttling_exception

See KmsThrottlingException

not_found_exception

See NotFoundException

opted_out_exception

See OptedOutException

platform_application_disabled_exception

See PlatformApplicationDisabledException

resource_not_found_exception

See ResourceNotFoundException

stale_tag_exception

See StaleTagException

subscription_limit_exceeded_exception

See SubscriptionLimitExceededException

tag_limit_exceeded_exception

See TagLimitExceededException

tag_policy_exception

See TagPolicyException

throttled_exception

See ThrottledException

topic_limit_exceeded_exception

See TopicLimitExceededException

user_error_exception

See UserErrorException

validation_exception

See ValidationException

verification_exception

See VerificationException

Structs

AddPermissionError

AuthorizationErrorException

Indicates that the user has been denied access to the requested resource.

CheckIfPhoneNumberIsOptedOutError

ConcurrentAccessException

Can't perform multiple operations on a tag simultaneously. Perform the operations

ConfirmSubscriptionError

CreatePlatformApplicationError

CreatePlatformEndpointError

CreateSMSSandboxPhoneNumberError

CreateTopicError

DeleteEndpointError

DeletePlatformApplicationError

DeleteSMSSandboxPhoneNumberError

DeleteTopicError

EndpointDisabledException

Exception error indicating endpoint disabled.

FilterPolicyLimitExceededException

Indicates that the number of filter polices in your AWS account exceeds the limit. To

GetEndpointAttributesError

GetPlatformApplicationAttributesError

GetSMSAttributesError

GetSMSSandboxAccountStatusError

GetSubscriptionAttributesError

GetTopicAttributesError

InternalErrorException

Indicates an internal service error.

InvalidParameterException

Indicates that a request parameter does not comply with the associated

InvalidParameterValueException

Indicates that a request parameter does not comply with the associated

InvalidSecurityException

The credential signature isn't valid. You must use an HTTPS endpoint and sign your

KmsAccessDeniedException

The ciphertext references a key that doesn't exist or that you don't have access

KmsDisabledException

The request was rejected because the specified customer master key (CMK) isn't

KmsInvalidStateException

The request was rejected because the state of the specified resource isn't valid for

KmsNotFoundException

The request was rejected because the specified entity or resource can't be

KmsOptInRequired

The AWS access key ID needs a subscription for the service.

KmsThrottlingException

The request was denied due to request throttling. For more information about

ListEndpointsByPlatformApplicationError

ListOriginationNumbersError

ListPhoneNumbersOptedOutError

ListPlatformApplicationsError

ListSMSSandboxPhoneNumbersError

ListSubscriptionsByTopicError

ListSubscriptionsError

ListTagsForResourceError

ListTopicsError

NotFoundException

Indicates that the requested resource does not exist.

OptInPhoneNumberError

OptedOutException

Indicates that the specified phone number opted out of receiving SMS messages from

PlatformApplicationDisabledException

Exception error indicating platform application disabled.

PublishError

RemovePermissionError

ResourceNotFoundException

Can’t perform the action on the specified resource. Make sure that the resource

SetEndpointAttributesError

SetPlatformApplicationAttributesError

SetSMSAttributesError

SetSubscriptionAttributesError

SetTopicAttributesError

StaleTagException

A tag has been added to a resource with the same ARN as a deleted resource. Wait a

SubscribeError

SubscriptionLimitExceededException

Indicates that the customer already owns the maximum allowed number of

TagLimitExceededException

Can't add more than 50 tags to a topic.

TagPolicyException

The request doesn't comply with the IAM tag policy. Correct your request and then

TagResourceError

ThrottledException

Indicates that the rate at which requests have been submitted for this action exceeds

TopicLimitExceededException

Indicates that the customer already owns the maximum allowed number of topics.

UnsubscribeError

UntagResourceError

UserErrorException

Indicates that a request parameter does not comply with the associated

ValidationException

Indicates that a parameter in the request is invalid.

VerificationException

Indicates that the one-time password (OTP) used for verification is invalid.

VerifySMSSandboxPhoneNumberError

Enums

AddPermissionErrorKind

CheckIfPhoneNumberIsOptedOutErrorKind

ConfirmSubscriptionErrorKind

CreatePlatformApplicationErrorKind

CreatePlatformEndpointErrorKind

CreateSMSSandboxPhoneNumberErrorKind

CreateTopicErrorKind

DeleteEndpointErrorKind

DeletePlatformApplicationErrorKind

DeleteSMSSandboxPhoneNumberErrorKind

DeleteTopicErrorKind

GetEndpointAttributesErrorKind

GetPlatformApplicationAttributesErrorKind

GetSMSAttributesErrorKind

GetSMSSandboxAccountStatusErrorKind

GetSubscriptionAttributesErrorKind

GetTopicAttributesErrorKind

ListEndpointsByPlatformApplicationErrorKind

ListOriginationNumbersErrorKind

ListPhoneNumbersOptedOutErrorKind

ListPlatformApplicationsErrorKind

ListSMSSandboxPhoneNumbersErrorKind

ListSubscriptionsByTopicErrorKind

ListSubscriptionsErrorKind

ListTagsForResourceErrorKind

ListTopicsErrorKind

OptInPhoneNumberErrorKind

PublishErrorKind

RemovePermissionErrorKind

SetEndpointAttributesErrorKind

SetPlatformApplicationAttributesErrorKind

SetSMSAttributesErrorKind

SetSubscriptionAttributesErrorKind

SetTopicAttributesErrorKind

SubscribeErrorKind

TagResourceErrorKind

UnsubscribeErrorKind

UntagResourceErrorKind

VerifySMSSandboxPhoneNumberErrorKind