Module aws_sdk_sns::model

Modules

endpoint

See Endpoint

message_attribute_value

See MessageAttributeValue

phone_number_information

See PhoneNumberInformation

platform_application

See PlatformApplication

sms_sandbox_phone_number

See SmsSandboxPhoneNumber

subscription

See Subscription

tag

See Tag

topic

See Topic

Structs

Endpoint

Endpoint for mobile app and device.

MessageAttributeValue

The user-specified message attribute value. For string data types, the value attribute

PhoneNumberInformation

A list of phone numbers and their metadata.

PlatformApplication

Platform application object.

SmsSandboxPhoneNumber

A verified or pending destination phone number in the SMS sandbox.

Subscription

A wrapper type for the attributes of an Amazon SNS subscription.

Tag

The list of tags to be added to the specified topic.

Topic

A wrapper type for the topic's Amazon Resource Name (ARN). To retrieve a topic's

Enums

LanguageCodeString

Supported language code for sending OTP message

NumberCapability

Enum listing out all supported number capabilities.

RouteType

Enum listing out all supported route types. The following enum values are supported.

SmsSandboxPhoneNumberVerificationStatus

Enum listing out all supported destination phone number verification statuses. The following enum values are supported.