Module aws_sdk_sns::operation

Structs

AddPermission

Adds a statement to a topic's access control policy, granting access for the specified

CheckIfPhoneNumberIsOptedOut

Accepts a phone number and indicates whether the phone holder has opted out of

ConfirmSubscription

Verifies an endpoint owner's intent to receive messages by validating the token sent

CreatePlatformApplication

Creates a platform application object for one of the supported push notification

CreatePlatformEndpoint

Creates an endpoint for a device and mobile app on one of the supported push

CreateSMSSandboxPhoneNumber

Adds a destination phone number to an AWS account in the SMS sandbox and sends a

CreateTopic

Creates a topic to which notifications can be published. Users can create at most

DeleteEndpoint

Deletes the endpoint for a device and mobile app from Amazon SNS. This action is

DeletePlatformApplication

Deletes a platform application object for one of the supported push notification

DeleteSMSSandboxPhoneNumber

Deletes an AWS account's verified or pending phone number from the SMS sandbox.

DeleteTopic

Deletes a topic and all its subscriptions. Deleting a topic might prevent some

GetEndpointAttributes

Retrieves the endpoint attributes for a device on one of the supported push

GetPlatformApplicationAttributes

Retrieves the attributes of the platform application object for the supported push

GetSMSAttributes

Returns the settings for sending SMS messages from your account.

GetSMSSandboxAccountStatus

Retrieves the SMS sandbox status for the calling AWS account in the target AWS

GetSubscriptionAttributes

Returns all of the properties of a subscription.

GetTopicAttributes

Returns all of the properties of a topic. Topic properties returned might differ based

ListEndpointsByPlatformApplication

Lists the endpoints and endpoint attributes for devices in a supported push

ListOriginationNumbers

Lists the calling AWS account's dedicated origination numbers and their metadata. For

ListPhoneNumbersOptedOut

Returns a list of phone numbers that are opted out, meaning you cannot send SMS

ListPlatformApplications

Lists the platform application objects for the supported push notification services,

ListSMSSandboxPhoneNumbers

Lists the calling AWS account's current verified and pending destination phone numbers

ListSubscriptions

Returns a list of the requester's subscriptions. Each call returns a limited list of

ListSubscriptionsByTopic

Returns a list of the subscriptions to a specific topic. Each call returns a limited

ListTagsForResource

List all tags added to the specified Amazon SNS topic. For an overview, see Amazon SNS Tags in the

ListTopics

Returns a list of the requester's topics. Each call returns a limited list of topics,

OptInPhoneNumber

Use this request to opt in a phone number that is opted out, which enables you to

Publish

Sends a message to an Amazon SNS topic, a text message (SMS message) directly to a phone

RemovePermission

Removes a statement from a topic's access control policy.

SetEndpointAttributes

Sets the attributes for an endpoint for a device on one of the supported push

SetPlatformApplicationAttributes

Sets the attributes of the platform application object for the supported push

SetSMSAttributes

Use this request to set the default settings for sending SMS messages and receiving

SetSubscriptionAttributes

Allows a subscription owner to set an attribute of the subscription to a new

SetTopicAttributes

Allows a topic owner to set an attribute of the topic to a new value.

Subscribe

Subscribes an endpoint to an Amazon SNS topic. If the endpoint type is HTTP/S or email, or

TagResource

Add tags to the specified Amazon SNS topic. For an overview, see Amazon SNS Tags in the

Unsubscribe

Deletes a subscription. If the subscription requires authentication for deletion, only

UntagResource

Remove tags from the specified Amazon SNS topic. For an overview, see Amazon SNS Tags in the

VerifySMSSandboxPhoneNumber

Verifies a destination phone number with a one-time password (OTP) for the calling AWS