Module aws_sdk_sns::error

Modules

authorization_error_error

See AuthorizationErrorError

concurrent_access_error

See ConcurrentAccessError

endpoint_disabled_error

See EndpointDisabledError

filter_policy_limit_exceeded_error

See FilterPolicyLimitExceededError

internal_error_error

See InternalErrorError

invalid_parameter_error

See InvalidParameterError

invalid_parameter_value_error

See InvalidParameterValueError

invalid_security_error

See InvalidSecurityError

kms_access_denied_error

See KMSAccessDeniedError

kms_disabled_error

See KMSDisabledError

kms_invalid_state_error

See KMSInvalidStateError

kms_not_found_error

See KMSNotFoundError

kms_opt_in_required

See KMSOptInRequired

kms_throttling_error

See KMSThrottlingError

not_found_error

See NotFoundError

opted_out_error

See OptedOutError

platform_application_disabled_error

See PlatformApplicationDisabledError

resource_not_found_error

See ResourceNotFoundError

stale_tag_error

See StaleTagError

subscription_limit_exceeded_error

See SubscriptionLimitExceededError

tag_limit_exceeded_error

See TagLimitExceededError

tag_policy_error

See TagPolicyError

throttled_error

See ThrottledError

topic_limit_exceeded_error

See TopicLimitExceededError

user_error_error

See UserErrorError

validation_error

See ValidationError

verification_error

See VerificationError

Structs

AddPermissionError

AuthorizationErrorError

Indicates that the user has been denied access to the requested resource.

CheckIfPhoneNumberIsOptedOutError

ConcurrentAccessError

Can't perform multiple operations on a tag simultaneously. Perform the operations

ConfirmSubscriptionError

CreatePlatformApplicationError

CreatePlatformEndpointError

CreateSMSSandboxPhoneNumberError

CreateTopicError

DeleteEndpointError

DeletePlatformApplicationError

DeleteSMSSandboxPhoneNumberError

DeleteTopicError

EndpointDisabledError

Exception error indicating endpoint disabled.

FilterPolicyLimitExceededError

Indicates that the number of filter polices in your AWS account exceeds the limit. To

GetEndpointAttributesError

GetPlatformApplicationAttributesError

GetSMSAttributesError

GetSMSSandboxAccountStatusError

GetSubscriptionAttributesError

GetTopicAttributesError

InternalErrorError

Indicates an internal service error.

InvalidParameterError

Indicates that a request parameter does not comply with the associated

InvalidParameterValueError

Indicates that a request parameter does not comply with the associated

InvalidSecurityError

The credential signature isn't valid. You must use an HTTPS endpoint and sign your

KMSAccessDeniedError

The ciphertext references a key that doesn't exist or that you don't have access

KMSDisabledError

The request was rejected because the specified customer master key (CMK) isn't

KMSInvalidStateError

The request was rejected because the state of the specified resource isn't valid for

KMSNotFoundError

The request was rejected because the specified entity or resource can't be

KMSOptInRequired

The AWS access key ID needs a subscription for the service.

KMSThrottlingError

The request was denied due to request throttling. For more information about

ListEndpointsByPlatformApplicationError

ListOriginationNumbersError

ListPhoneNumbersOptedOutError

ListPlatformApplicationsError

ListSMSSandboxPhoneNumbersError

ListSubscriptionsByTopicError

ListSubscriptionsError

ListTagsForResourceError

ListTopicsError

NotFoundError

Indicates that the requested resource does not exist.

OptInPhoneNumberError

OptedOutError

Indicates that the specified phone number opted out of receiving SMS messages from

PlatformApplicationDisabledError

Exception error indicating platform application disabled.

PublishError

RemovePermissionError

ResourceNotFoundError

Can’t perform the action on the specified resource. Make sure that the resource

SetEndpointAttributesError

SetPlatformApplicationAttributesError

SetSMSAttributesError

SetSubscriptionAttributesError

SetTopicAttributesError

StaleTagError

A tag has been added to a resource with the same ARN as a deleted resource. Wait a

SubscribeError

SubscriptionLimitExceededError

Indicates that the customer already owns the maximum allowed number of

TagLimitExceededError

Can't add more than 50 tags to a topic.

TagPolicyError

The request doesn't comply with the IAM tag policy. Correct your request and then

TagResourceError

ThrottledError

Indicates that the rate at which requests have been submitted for this action exceeds

TopicLimitExceededError

Indicates that the customer already owns the maximum allowed number of topics.

UnsubscribeError

UntagResourceError

UserErrorError

Indicates that a request parameter does not comply with the associated

ValidationError

Indicates that a parameter in the request is invalid.

VerificationError

Indicates that the one-time password (OTP) used for verification is invalid.

VerifySMSSandboxPhoneNumberError

Enums

AddPermissionErrorKind

CheckIfPhoneNumberIsOptedOutErrorKind

ConfirmSubscriptionErrorKind

CreatePlatformApplicationErrorKind

CreatePlatformEndpointErrorKind

CreateSMSSandboxPhoneNumberErrorKind

CreateTopicErrorKind

DeleteEndpointErrorKind

DeletePlatformApplicationErrorKind

DeleteSMSSandboxPhoneNumberErrorKind

DeleteTopicErrorKind

GetEndpointAttributesErrorKind

GetPlatformApplicationAttributesErrorKind

GetSMSAttributesErrorKind

GetSMSSandboxAccountStatusErrorKind

GetSubscriptionAttributesErrorKind

GetTopicAttributesErrorKind

ListEndpointsByPlatformApplicationErrorKind

ListOriginationNumbersErrorKind

ListPhoneNumbersOptedOutErrorKind

ListPlatformApplicationsErrorKind

ListSMSSandboxPhoneNumbersErrorKind

ListSubscriptionsByTopicErrorKind

ListSubscriptionsErrorKind

ListTagsForResourceErrorKind

ListTopicsErrorKind

OptInPhoneNumberErrorKind

PublishErrorKind

RemovePermissionErrorKind

SetEndpointAttributesErrorKind

SetPlatformApplicationAttributesErrorKind

SetSMSAttributesErrorKind

SetSubscriptionAttributesErrorKind

SetTopicAttributesErrorKind

SubscribeErrorKind

TagResourceErrorKind

UnsubscribeErrorKind

UntagResourceErrorKind

VerifySMSSandboxPhoneNumberErrorKind